# docs/swagger.yaml
basePath: /
definitions:
...
  services.GetMedicalLabScientistProfessionalDetailsResponseDto:
    properties:
      data:
        $ref: '#/definitions/services.MedicalLabScientistProfessionalDetailsRes'
      message:
        type: string
      status:
        type: boolean
    type: object
...
  services.MedicalLabScientistProfessionalDetailsRes:
    properties:
      department:
        type: string
      professionalCertificate:
        type: string
      professionalLicense:
        type: string
      qualification:
        type: string
    type: object
...
  services.UpdateMedicalLabScientistProfessionalInfoReqDto:
    properties:
      department:
        type: string
      qualifications:
        type: string
    type: object
  services.UpdateMedicalLabScientistProfessionalInfoResDto:
    properties:
      message:
        type: string
      status:
        type: boolean
    type: object
...
host: mmpn2atpcm.eu-west-1.awsapprunner.com
info:
  contact:
    name: Chirag Sharma
  description: Care Ville Backend (in GoLang)
  license:
    name: MIT
  title: Care Ville
  version: 0.0.1
paths:
...
  /provider/services/edit-medicalLabScientist-professional-info:
    put:
      consumes:
      - multipart/form-data
      description: Update provider
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update data of provider
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/services.UpdateMedicalLabScientistProfessionalInfoReqDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.UpdateMedicalLabScientistProfessionalInfoResDto'
      summary: Update provider
      tags:
      - medicalLabScientist
...
  /provider/services/get-medicalLabScientist-professional-details:
    get:
      consumes:
      - application/json
      description: Fetch professionalDetails By ID
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.GetMedicalLabScientistProfessionalDetailsResponseDto'
      summary: Fetch professionalDetails By ID
      tags:
      - medicalLabScientist
...
swagger: "2.0"